requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
get==2019.4.13
idna==2.8
lxml==4.4.2
pkg-resources==0.0.0
post==2019.4.13
public==2019.4.13
pycodestyle==2.5.0
pycparser==2.19
python-telegram-bot==12.2.0
query-string==2019.4.13
redis==3.3.11
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
tornado==6.0.3
urllib3==1.25.7
